Refactor ServiceEntityRepository constructor and add save and delete methods
This commit refactors the constructor of the ServiceEntityRepository class to use a more descriptive parameter name and adds two new methods: save and delete. These methods handle the persistence and deletion of entities respectively. Additionally, a private method validateEntityType is introduced to ensure that the given entity object is of the expected type.